Made a basic TUI.
Decided to switch to a custom sidebar from the bubbles list component. You can cycle through channels in the sidebar but they are just placeholders and you can't open a channel for now.
You can focus the component of the TUI you want to interact with using tab or shift+tab. The focused component has different border color.
Learnt about websocket. Using it to display messages sent in slack in the currently open channel. The messages don't appear for some reason, will have to fix that